+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **Breaking**: `aviutl2::config::translate`の戻り値を`Result<String>`から`String`に変更
 - eframe: persistence feature flagを追加
 - generic: デフォルトでエラー時にビープ音を鳴らすように
+- generic: `EditSection::objects_in_layer`を追加
 
 ### デモプラグイン